extracting auth functions into seperate lib folder
